refactor: get javascript file just once, move to utils
